third_party/blink/web_tests/external/wpt/html/semantics/scripting-1/the-script-element/execution-timing/106-external-module-import-expected.txt

This is a testharness.js-based test.
FAIL  scheduler: stylesheets blocking external module scripts assert_equals: expected "fixed" but got "static"
Harness: the test ran to completion.
